Implement guarded lenses in datum plans
Using the environment variable DATUM_PLAN_GUARDS (or manipulating the
`guardedGroups` property of the primary export of
`natural-lenses/datum-plan`), one or more datum plans can be selected to
receive special attention when undefined properties are accessed.  This
typically occurs when someone mistypes a property name in a program and
can be extremely difficult to trace in Javascript if the error occurs at
the last property access because of the resulting value `undefined`
instead of an error.

If datum plan guards are activated for a datum plan, Javascript Proxies
are wrapped around the lenses such that any undefined property actually
returns a Proxied object that throws an error when any of its properties
are accessed, but cites the original undefined property access in the
error trace.

Issue: #5