Add pre-checks for `npm publish`
Testing that the working directory is "hygienic" before publishing means
to things:
- There are no modifications between HEAD and the working tree.
- HEAD is on the appropriate version branch at `origin`.

The first requirement ensure that no-one publishes code changes not
recorded in Git.  The second requirement ensures that someone else was
able to see the commit which was published *and* that publication is
taking place from the intended branch.